rag/requirements.txt
langchain==1.2.10
langchain-openai==1.1.10
langchain-community==0.4.1
openai==2.26.0
faiss-cpu==1.13.2
pypdf==6.7.5
python-dotenv==1.2.2
requests==2.32.5
